Fix babel config for webpack example
preset-env now emits es modules by default and needs to be told explicitly
not to for the loader to work.

TODO: I think it wouldn't be impossible to have the loader work with ES6
modules, but that work is out of the scope of this PR. Might do it later
if there's a real use case for it. Note, though, that it's pretty easy to
configure babel to use commonjs output only in dev (where this loader is
used) and ES6 modules in prod (to get webpack features like tree shaking).