Proxy properties from $templateRequest to the mock
Fixes #14.
We override angular's $templateRequest service for templates' hot
reloading, but the mock service didn't provide access to the properties
attached to the original service function. ngAnimate uses the property
`$templateRequest.totalPendingRequests`, which wasn't available in the
mock.

Create simple getters and setters that work as a proxy to the original
service's own properties.