Fix accidental assignment in conditional
Turns out the standalone package has had a rather embarrassing bug this
whole time; readyState check in if was accidentally assignment (=)
instead of an equality check (===). Apparently things still worked ok
this long... it only started throwing now as apparently newer versions
of the ws library declare this property as readonly.

Fix the bug and add an eslint rule that should prevent similar mistakes
in the future.