Move lodash yarn resolution to top level package.json
We force peer dependency browsersync->lodash->opt-merger->lodash to use
version 4.17.21 instead of an old version 3.x that the opt-merger peer
dependency declares for itself.

This keeps npm audit happy, as there are a few CVE:s for the older
versions of lodash. We already did this in demo package's own
package.json, but apparently it's now needed in the top level
package.json, perhaps due to a change in Lerna upgrade?