Disable LF line ending requirement
It hasn't been properly enforced this far and now most of the files do
have windows line endings. I'd love to just move to fully LF but that
would modify all of the lines in all of the files and I don't think it's
worth messing up git blame and all.